backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
comm==0.2.2
contourpy==1.2.0
cryptography==45.0.5
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.1.0
fastapi==0.115.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fonttools==4.47.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
nodemon==0.0.17
numpy==1.26.3
openai==1.70.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.1
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic-settings==2.6.0
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.12.0
selenium==4.27.1
sentry-sdk==2.35.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.19
stack-data==0.6.3
starlette==0.38.6
sympy==1.12
threadpoolctl==3.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typer==0.16.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
